Separate the tools- from the testSuite-classpath in the bootstrap test.
Previously, the classpaths of the tools and the testSuite were
merged and mixed together in the bootstrap test:

* toolsJS/test:fullClasspath contained both, *and*
* the classpath optimized by the bootstrapped optimizer contained
  both.

This commit cleanly distinguishes the two. Now,
`toolsJS/test:fullClasspath`, which is optimized by the JVM linker,
only contains the tools, with `QuickLinker` as the entry point.
The classpath given to `QuickLinker`, and hence optimized by the
bootstrapped linker, only contains the test suite, including a
`ConsoleTestRunner` as entry point.
